requirements.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
dominate==2.8.0
elastic-transport==8.4.0
elasticsearch==8.8.2
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.0
PyJWT==2.7.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
urllib3==1.26.16
validators==0.20.0
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
WTForms-Components==0.10.5
